refactor(chat): extract chat_manager/construct.rs (#292)
Move ChatManager::new + its lifecycle plumbing into a dedicated
submodule and shrink the constructor to a Collaborators::build call so
it fits the 50-line function limit.
